Refactor tests to use common fixtures
