+++ tests/components/collection_image/const.py
@@ -1,0 +1,6 @@
+"""Constants for the Collection Image integration tests."""
+
+from pathlib import Path
+
+TEST_IMAGE = Path(__file__).parent / "test.png"
+DEFAULT_ENTITY_ID = "image.random_image"
